requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Pillow==10.0.0
Werkzeug==2.3.7
bcrypt==4.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.20
email-validator==2.0.0
pytz==2023.3
Flask-WTF==1.1.1
pytest==7.4.0
pytest-flask==1.2.0
Flask-Limiter==3.5.0